pyproject.toml
[tool.poetry]
name = "ddobok-backend"
version = "0.1.0"
description = ""
authors = ["eakyoungyou <eakyoungyu@gmail.com>"]
license = "MIT"
readme = "README.md"
packages = [{include = "ddobok_backend"}]

[tool.poetry.dependencies]
python = "3.9.5"
django = "~4.0"
flake8 = "^6.0.0"
black = "^22.12.0"
pillow = "^9.4.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^3.13.0"
dj-database-url = "^1.2.0"
psycopg2-binary = "^2.9.5"
whitenoise = {extras = ["brotli"], version = "^6.3.0"}
gunicorn = "^20.1.0"
boto3 = "^1.26.56"
django-storages = "^1.13.2"
django-filter = "^22.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
